Rename some packages for clarity
