+++ azure/virtualmachinedisk/client.go
@@ -1,4 +1,4 @@
-package disk
+package virtualmachinedisk
 
 import (
 	"fmt"
